+++ esbuild.config.mjs
@@ -32,6 +32,7 @@ const context = await esbuild.context({
 	],
 	format: 'cjs',
 	target: 'es2021',
+	define: { global: 'globalThis' },
 	logLevel: 'info',
 	sourcemap: prod ? false : 'inline',
 	treeShaking: true,